Clean Up and Final Code
I think this is correct for all the code we need for the final code, except for Aadi matlab app. It shouldn't be hard to implement what we have here with that. I also cleaned up some of the files to make it slightly easier to navigate and I commented a lot of the functions.